\name{bday}
\alias{bday}

\title{Custom gift generator for birthdays}

\description{
  This function will automatically generate gifts and presents for birthday
  boys or girls.
}

\details{
  All you have to do is add your name to this function as it is the only
  parameter.
}

\usage{
  bday(name = NULL)
}
\arguments{
  \item{name}{Your name or whatever character string you want. I don't care.}
}

\author{Brandon Monier, \email{brandon.monier@sdstate.edu} }

\examples{
  \code{
    bday(name = "Ronald Reagan")
  }
}

\keyword{birthdays}